Extracted NetworkRunner in separate file (with Graphz temp code)
